qcom-cpufreq: Add underclock for perf and power clusters
Underclock can be disabled by passing 'no_underclock' to the kernel command
line.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

qcom-cpufreq: Increase power cluster max freq to 1401 MHz

CPU0 is part of the power cluster and it's never unplugged (it's the
primary CPU), so the extra juice is needed.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

qcom-cpufreq: Raise underclocked freqs and fix no-underclock switch

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

qcom-cpufreq: Reduce freq limits (again)

*Back to how they were before 85eb82d6caf64e83b49b72d90364139ecad8f1df

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

qcom-cpufreq: Raise freq limit for both clusters

Perf cluster: 1920 MHz
Power cluster: 1478 MHz

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

qcom-cpufreq: Add underclocked frequencies for OnePlus 3T (MSM8996pro)

Performance cluster max: 2054400 KHz
Power cluster max: 1593600 KHz

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>

qcom-cpufreq: Update underclock for MSM8996 and MSM8996pro

Apparently, the higher stock frequencies on both MSM8996 and MSM8996pro are
unstable, even with a kernel binary that the OEM created. Running AnTuTu
with thermal throttling disabled (to force the CPUs to remain at their
respective maxfreqs) _always_ crashes MSM8996pro at its stock maxfreqs, and
occasionally crashes MSM8996 at its stock maxfreqs. A workaround to this is
to underclock the power cluster (pwrcl) on both chipsets.

Underclocking the pwrcl to 1516 MHz on MSM8996pro makes the performance
cluster (perfcl) stable up to 2150 MHz; without the pwrcl underclock, the
AnTuTu crashes could be reproduced 100% of the time on MSM8996pro even when
the perfcl was reduced to 1824 MHz. This underclock does not cure stability
at the perfcl's stock maxfreq of 2342 MHz though, so do not allow the
underclock to be disabled at all for both clusters on MSM8996pro.

On MSM8996, the perfcl's stock maxfreq of 2150 MHz is stable when the pwrcl
is underclocked to 1478 MHz. Therefore, allow the underclock to be disabled
_only_ for the perfcl on MSM8996 (pwrcl must always remain underclocked).

This is a very hacky solution to a very serious bug, so add TODOs to
investigate this further. Kernel logs are clean for these crashes, and
they could still be reproduced with a 20W wall charger connected to the
device, so they are not caused by a battery issue.

Change-Id: I848e50857ac187321ebf20474d48306b7451fb2d
Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>