ARM: dts: 15801: Disable core hotplugging, up the mitigation freq
Don't throttle as much, don't hotplug cores. There are two potential
cases:

1) The screen is off and the device is in deep sleep. The cores will
actually be disabled as the device enters a sleep state so the BCL is
useless.

2) The user is actively using the device. We hotplug out the third core
and throttle the rest to ~500mhz. While this may seem a good idea, the
device actually becomes unusable and takes longer to complete tasks and
go back to low power states, which actually makes the problem worse.

This patch disables hotplugging and ups the throttle freq to 1132800
which makes the phone usable while skipping the higher freq states.

Change-Id: I07ca3c62d53c65f4197ce6627f29f722d39f2c28

ARM: dts: 15801: Decommonize BCL mitigation freq

 * 1132800 kHz is only available on msm8996pro

Change-Id: I47d98a167b01ded846a45bfa296d3b9c4b5acdb5